requirements.txt
torch
torchvision
gradio
datasets
matplotlib
numpy
Pillow
torchinfo